refactor: 인터뷰어 카드, 선택 버튼, 로직 및 클라이언트 컴포넌트로 분리
