+++ apps/frontend/src/app/interviewer/_components/InterviewerCard/index.tsx
@@ -1,11 +1,8 @@
-"use client";
-
 import Image from "next/image";
 import styles from "./InterviewerCard.module.css";
 import Text from "@repo/ui/Text";
-import Button from "@repo/ui/Button";
-import { useInterviewerStore } from "@/store/useInterviewerStore";
-import { useRouter } from "next/navigation";
+import StartInterviewButton from "./StartInterviewButton";
+
 interface InterviewerCardProps {
   id: number;
   imgUrl: string;
@@ -21,21 +18,6 @@ export default function InterviewerCard({
   mbti,
   description,
 }: InterviewerCardProps): React.ReactElement {
-  const setInterviewer = useInterviewerStore((state) => state.setInterviewer);
-  const router = useRouter();
-  const handleClick = () => {
-    setInterviewer({
-      id,
-      name,
-      imgUrl,
-      mbti,
-      description,
-    });
-    router.push(`/chat/31`);
-  };
-
-  console.log(id, name, imgUrl, mbti, description);
-
   return (
     <div className={styles.wrapper}>
       <Image src={imgUrl} alt={name} height={150} width={150} />
@@ -50,9 +32,7 @@ export default function InterviewerCard({
           {mbti}
         </Text>
         <div className={styles.buttonWrapper}>
-          <Button variant="outline" size="md" onClick={handleClick}>
-            Start interview
-          </Button>
+          <StartInterviewButton id={id} />
         </div>
       </div>
     </div>
